move format bytes to util
