+++ src/utils/ObjectUtils.ts
@@ -1,0 +1,13 @@
+export const formatBytes = (size: number | undefined): string => {
+  if (!size || size == 0) {
+    return '-';
+  }
+
+  const units = ['B', 'KB', 'MB', 'GB', 'TB'];
+  let n = 0;
+  while (size >= 1024 && n < units.length - 1) {
+    size /= 1024;
+    n++;
+  }
+  return `${size.toFixed(2)} ${units[n]}`;
+};
